requirements.txt

#### 2. `requirements.txt`

```txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pyyaml>=6.0
tenacity>=8.2.3
python-git>=0.2
age>=0.2.0
streamlit>=1.29.0
pypowerwall>=0.5.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-prometheus>=0.42b0